Fixed the Nullable Error as well as some method changes coinciding. Problem with ArrayList present in the loops for populateData says accessing index=70 while size =70

+++ app/src/main/java/com/example/textmate/alg.java
@@ -41,7 +41,7 @@ public class alg{
     private double avgTimeRec,avgTimeSent,WordPerMessage;
     public DatabaseHelper upDBScores;
     //Constructors
-    public alg(int ID,int numTxt,int txtNumWord,double txtTimeSent,double txtTimeReceived,double wordPerMessage,double valYesterday,double valAvg,double currScore, int numberUpdate){
+    public alg(Long ID,int numTxt,int txtNumWord,double txtTimeSent,double txtTimeReceived,double wordPerMessage,double valYesterday,double valAvg,double currScore, int numberUpdate){
         this.WordPerMessage = wordPerMessage;
         this.wordCount = numTxt;
         this.numUpdate = numberUpdate;
